third_party/WebKit/LayoutTests/virtual/threaded/fast/compositorworker/compositor-proxy-supports.html

<!DOCTYPE html>
<meta charset="utf-8">
<style>
#container {
  width: 100px;
  height: 100px;
  border: 1px solid black;
  overflow: scroll;
}

#scrolled {
  width: 80px;
  height: 200px;
  background: blue;
}
</style>
<div id="container">
  <div id="scrolled"></div>
</div>
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<script>
test(function() {
    var proxy = new CompositorProxy(document.getElementById('container'), ['opacity', 'transform', 'scrollTop', 'scrollLeft']);
    assert_true(proxy.supports('opacity'));
    assert_false(proxy.supports('opacit'));
    assert_false(proxy.supports('opacitya'));

    assert_true(proxy.supports('transform'));
    assert_false(proxy.supports('transfor'));
    assert_false(proxy.supports('transforma'));

    assert_true(proxy.supports('scrollLeft'));
    assert_false(proxy.supports('scrollLef'));
    assert_false(proxy.supports('scrollLefta'));

    assert_true(proxy.supports('scrollTop'));
    assert_false(proxy.supports('scrollTo'));
    assert_false(proxy.supports('scrollTopa'));

    assert_false(proxy.supports('あ'));

    // This is a hack to get a 16-bit string for a supported property.
    var encoded = encode_utf16('transform', true);
    var transform16 = new TextDecoder('utf-16').decode(encoded);
    assert_true(transform16 === 'transform');
    assert_true(proxy.supports(transform16));
}, "This test ensures that supports correctly handles queries with strings that a prefix of a property or vice versa.");

function encode_utf16(s, littleEndian) {
    var a = new Uint8Array(s.length * 2), view = new DataView(a.buffer);
    s.split('').forEach(function(c, i) {
        view.setUint16(i * 2, c.charCodeAt(0), littleEndian);
    });
    return a;
}
</script>
